requirements.txt
# requirements.txt
requests==2.31.0
gspread==5.12.4
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
six==1.16.0
cachetools==5.3.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
rsa==4.9

# For local development and testing
pytest==7.4.3
pytest-mock==3.12.0
responses==0.24.1
python-decouple==3.8


